Split _resolvers.py into a package mirroring _plugins/
Each builtin resolver moves to its own submodule (``venv.py``,
``pyproject.py``, ``uv_workspace.py``); the protocol, ``PathMap``
type, and ``merge_paths`` utility live in ``_core.py``; ``exported_roots``
gets its own ``_exports.py``. The package ``__init__.py`` re-exports
everything that was importable from ``_resolvers.py`` so external
imports keep working unchanged.

No behavior change.
